Update update_resources.sh
只在wget返回退出码0时判定文件下载成功（否则，有概率出现文件损坏导致后续防火墙规则有问题）。

+++ root/etc/homeproxy/scripts/update_resources.sh
@@ -68,7 +68,7 @@ check_list_update() {
 	fi
 
 	$wget "https://fastly.jsdelivr.net/gh/$listrepo@$list_sha/$listname" -O "$RUN_DIR/$listname"
-	if [ ! -s "$RUN_DIR/$listname" ]; then
+	if [ $? -eq 0 ] && [ ! -s "$RUN_DIR/$listname" ]; then
 		rm -f "$RUN_DIR/$listname"
 		log "[$(to_upper "$listtype")] Update failed."
